+++ e2e/evm/test/native_transfer.test.ts
@@ -26,7 +26,7 @@ describe("native transfer", () => {
 
     // Assert balances with logging
     const tenPow12 = toBigInt(1e12)
-    const gasUsed = 50000n // 50k gas for the transaction
+    const gasUsed = transaction.gasLimit
     const txCostMicronibi = amountToSend / tenPow12 + gasUsed
     const txCostWei = txCostMicronibi * tenPow12
     const expectedSenderWei = senderBalanceBefore - txCostWei
@@ -35,8 +35,9 @@ describe("native transfer", () => {
       amountToSend,
       expectedSenderWei,
       senderBalanceAfter,
+      txResponse,
     })
-    expect(senderBalanceAfter).toEqual(expectedSenderWei)
     expect(recipientBalanceAfter).toEqual(amountToSend)
+    expect(senderBalanceAfter).toEqual(expectedSenderWei)
   }, 20e3)
 })