+++ x/evm/tx_data_dynamic_fee.go
@@ -291,9 +291,9 @@ func (tx DynamicFeeTx) Cost() *big.Int {
 	return cost(tx.Fee(), tx.GetValueWei())
 }
 
-// EffectiveGasPriceWei returns the effective gas price based on EIP-1559 rules.
+// EffectiveGasPriceWeiPerGas returns the effective gas price based on EIP-1559 rules.
 // `effectiveGasPrice = min(baseFee + tipCap, feeCap)`
-func (tx *DynamicFeeTx) EffectiveGasPriceWei(baseFeeWei *big.Int) *big.Int {
+func (tx *DynamicFeeTx) EffectiveGasPriceWeiPerGas(baseFeeWei *big.Int) *big.Int {
 	feeWithSpecifiedTip := new(big.Int).Add(tx.GasTipCap.BigInt(), baseFeeWei)
 
 	// Enforce base fee as the minimum [EffectiveGasPriceWei]:
@@ -303,7 +303,7 @@ func (tx *DynamicFeeTx) EffectiveGasPriceWei(baseFeeWei *big.Int) *big.Int {
 
 // EffectiveFeeWei returns effective_gasprice * gaslimit.
 func (tx DynamicFeeTx) EffectiveFeeWei(baseFeeWei *big.Int) *big.Int {
-	return priceTimesGas(tx.EffectiveGasPriceWei(baseFeeWei), tx.GasLimit)
+	return priceTimesGas(tx.EffectiveGasPriceWeiPerGas(baseFeeWei), tx.GasLimit)
 }
 
 // EffectiveCost returns amount + effective_gasprice * gaslimit.