+++ x/evm/tx_data_legacy.go
@@ -190,14 +190,14 @@ func (tx LegacyTx) Cost() *big.Int {
 	return cost(tx.Fee(), tx.GetValueWei())
 }
 
-// EffectiveGasPriceWei is the same as GasPrice for LegacyTx
-func (tx LegacyTx) EffectiveGasPriceWei(baseFeeWei *big.Int) *big.Int {
+// EffectiveGasPriceWeiPerGas is the same as GasPrice for LegacyTx
+func (tx LegacyTx) EffectiveGasPriceWeiPerGas(baseFeeWei *big.Int) *big.Int {
 	return BigIntMax(tx.GetGasPrice(), baseFeeWei)
 }
 
 // EffectiveFeeWei is the same as Fee for LegacyTx
 func (tx LegacyTx) EffectiveFeeWei(baseFeeWei *big.Int) *big.Int {
-	return priceTimesGas(tx.EffectiveGasPriceWei(baseFeeWei), tx.GetGas())
+	return priceTimesGas(tx.EffectiveGasPriceWeiPerGas(baseFeeWei), tx.GetGas())
 }
 
 // EffectiveCost is the same as Cost for LegacyTx